feat:Corrige layout.tsx e estrutura do projeto para rodar corretamente
