+++ app/src/components/Header/Header.tsx
@@ -7,13 +7,6 @@ export default function Header() {
         </div>
         <h1 className="text-white">Sistema APAE</h1>
       </div>
-    <header className="p-4 bg-blue-600 text-white text-center flex items-center justify-center gap-3">
-      <img 
-        src="/logo.APAE.jpg" 
-        alt="APAE" 
-        className="w-16 h-16 md:w-18 md:h-18" 
-      />
-      <h1 className="text-lg md:text-xl font-bold">Gestão Escolar</h1>
     </header>
   );
 }
